Re-use FastSet library encoding/decoding
