+++ src/Tokenizer/Decompounder/Dictionary/FastSetDictionary.php
@@ -5,14 +5,35 @@ declare(strict_types=1);
 namespace Loupe\Matcher\Tokenizer\Decompounder\Dictionary;
 
 use Loupe\Matcher\Locale;
+use Toflar\FastSet\FastSet;
 
-class FastSetDictionary extends AbstractFastSetDictionary
+class FastSetDictionary implements DictionaryInterface
 {
-    public static function create(Locale $locale, string $directory): self
-    {
-        $self = new self($locale);
-        $self->loadFromDirectory($directory);
+    private const FILE_NAME_TERMS = 'terms';
 
-        return $self;
+    private FastSet $fastSet;
+
+    public function __construct(
+        protected Locale $locale,
+        string $directory
+    ) {
+        $this->fastSet = new FastSet($directory);
+
+        try {
+            $this->fastSet->initialize();
+        } catch (\Throwable) {
+            $this->fastSet->build($directory . '/' . self::FILE_NAME_TERMS);
+            $this->fastSet->initialize();
+        }
+    }
+
+    public function getLocale(): Locale
+    {
+        return $this->locale;
+    }
+
+    public function has(string $term): bool
+    {
+        return $this->fastSet->has($term);
     }
 }
